+++ src/App.tsx
@@ -1,35 +1,10 @@
-import React, { useMemo, useState } from 'react';
+import React, { useState } from 'react';
 import './App.scss';
-import { peopleFromServer } from './data/people';
-import debounce from 'lodash.debounce';
 import { Person } from './types/Person';
+import { Autocomplete } from './components/Autocomplete';
 
 export const App: React.FC = () => {
-  const [query, setQuery] = useState('');
-  const [appliedQuery, setAppliedQuery] = useState('');
-  const [showList, setShowList] = useState(false);
   const [selectedPerson, setSelectedPerson] = useState<Person | null>(null);
-
-  const filteredPeople = useMemo(() => {
-    return peopleFromServer.filter(person =>
-      person.name.toLowerCase().includes(appliedQuery.toLowerCase()),
-    );
-  }, [appliedQuery]);
-
-  const applyQuery = useMemo(
-    () => debounce((value: string) => setAppliedQuery(value.trim()), 300),
-    [],
-  );
-
-  const handleQueryChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-    setQuery(event.target.value);
-
-    if (selectedPerson && event.target.value !== selectedPerson.name) {
-      setSelectedPerson(null);
-    }
-
-    applyQuery(event.target.value);
-  };
 
   return (
     <div className="container">
@@ -40,61 +15,11 @@ export const App: React.FC = () => {
             : 'No selected person'}
         </h1>
 
-        <div className="dropdown is-active">
-          <div className="dropdown-trigger">
-            <input
-              type="text"
-              onChange={handleQueryChange}
-              onFocus={() => setShowList(true)}
-              onBlur={() => setShowList(false)}
-              value={query}
-              placeholder="Enter a part of the name"
-              className="input"
-              data-cy="search-input"
-            />
-          </div>
-
-          {showList && filteredPeople.length !== 0 && (
-            <div
-              className="dropdown-menu"
-              role="menu"
-              data-cy="suggestions-list"
-            >
-              <div className="dropdown-content">
-                {filteredPeople.map(person => (
-                  <div
-                    className="dropdown-item"
-                    data-cy="suggestion-item"
-                    key={person.name}
-                    onMouseDown={() => {
-                      setSelectedPerson(person);
-                      setQuery(person.name);
-                      setShowList(false);
-                    }}
-                  >
-                    <p className="has-text-link">{person.name}</p>
-                  </div>
-                ))}
-              </div>
-            </div>
-          )}
-        </div>
-
-        {showList && filteredPeople.length === 0 && (
-          <div
-            className="
-            notification
-            is-danger
-            is-light
-            mt-3
-            is-align-self-flex-start
-          "
-            role="alert"
-            data-cy="no-suggestions-message"
-          >
-            <p className="has-text-danger">No matching suggestions</p>
-          </div>
-        )}
+        <Autocomplete
+          onSelected={setSelectedPerson}
+          timeout={300}
+          selectedPerson={selectedPerson}
+        />
       </main>
     </div>
   );
